back/pyproject.toml
[project]
name = "ai-doc-translator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deep-translator>=1.11.4",
    "google-genai>=1.68.0",
    "anthropic>=0.25.0",
    "bcrypt>=4.2.0",
    "httpx>=0.27.0",
    "groq>=1.1.1",
    "openai>=2.29.0",
    "pandas>=2.3.3",
    "pikepdf>=10.5.1",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "python-pptx>=1.0.2",
    "seaborn>=0.13.2",
    "tqdm>=4.67.3",
    "watchdog>=6.0.0",
    "fastapi>=0.136.1",
    "uvicorn>=0.46.0",
    "python-multipart>=0.0.28",
    "websockets>=16.0",
    "pyjwt>=2.12.1",
    "authlib>=1.3.0",
    "itsdangerous>=2.2.0",
]
